requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.1
django-allauth==0.42.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-filter==2.3.0
django-js-asset==1.2.2
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-widget-tweaks==1.4.8
idna==2.10
jsonfield==3.1.0
oauthlib==3.1.0
psycopg2-binary==2.8.5
pycodestyle==2.6.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
swapper==1.1.2.post1
urllib3==1.25.9
